Change how test_documents mocks RPC calls
