+++ tests/profiles-enabled/test_shipments.py
@@ -2,7 +2,7 @@ import copy
 import json
 import re
 import random
-from unittest import mock
+from unittest import mock, TestCase
 
 import boto3
 import httpretty
@@ -26,7 +26,8 @@ from apps.shipments.rpc import Load110RPCClient, ShipmentRPCClient
 from apps.utils import random_id
 from apps.rpc_client import RPCClient, RPCError, requests
 from tests.utils import get_jwt
-from tests.utils import replace_variables_in_string, create_form_content, mocked_rpc_response
+from tests.utils import replace_variables_in_string, create_form_content, mocked_rpc_response, random_timestamp, \
+    random_location, GeoCoderResponse
 
 boto3.setup_default_session()  # https://github.com/spulec/moto/issues/1926
 
@@ -2403,411 +2404,3 @@ class DevicesLocationsAPITests(APITestCase):
             in_bbox_url = f'{url}?in_bbox=82.5,34.5,-82,45'
             response = self.client.get(in_bbox_url)
             self.assertEqual(response.status_code, status.HTTP_400_BAD_REQUEST)
-
-
-class TestShipmentRPCClient(APITestCase):
-    def test_vault_create(self):
-
-        rpc_client = ShipmentRPCClient()
-
-        # Error response from RPC Server should return server detail in exception
-        with mock.patch.object(requests.Session, 'post') as mock_method:
-            mock_method.return_value = mocked_rpc_response({
-                "error": {
-                    "code": 1337,
-                    "message": "Error from RPC Server",
-                },
-            })
-            try:
-                rpc_client.create_vault(STORAGE_CRED_ID, SHIPPER_WALLET_ID, CARRIER_WALLET_ID)
-                self.fail("Should have thrown RPC Error")
-            except RPCError as rpc_error:
-                self.assertEqual(rpc_error.status_code, 500)
-                self.assertEqual(rpc_error.detail, 'Error from RPC Server')
-
-            mock_method.return_value = mocked_rpc_response({
-                "result": {
-                    "code": 1337,
-                    "message": "Invalid response from Engine",
-                },
-            })
-            try:
-                rpc_client.create_vault(STORAGE_CRED_ID, SHIPPER_WALLET_ID, CARRIER_WALLET_ID)
-                self.fail("Should have thrown RPC Error")
-            except RPCError as rpc_error:
-                self.assertEqual(rpc_error.status_code, 500)
-                self.assertEqual(rpc_error.detail, 'Invalid response from Engine')
-
-            mock_method.return_value = mocked_rpc_response({
-                "jsonrpc": "2.0",
-                "result": {
-                    "success": True,
-                    "vault_id": CARRIER_WALLET_ID,
-                    "vault_uri": SHIPPER_WALLET_ID
-                },
-                "id": 0
-            })
-
-            vault_id, vault_uri = rpc_client.create_vault(STORAGE_CRED_ID, SHIPPER_WALLET_ID, CARRIER_WALLET_ID)
-            self.assertEqual(vault_id, CARRIER_WALLET_ID)
-            self.assertEqual(vault_uri, SHIPPER_WALLET_ID)
-
-    def test_add_shipment_data(self):
-
-        rpc_client = ShipmentRPCClient()
-
-        # Error response from RPC Server should return server detail in exception
-        with mock.patch.object(requests.Session, 'post') as mock_method:
-            mock_method.return_value = mocked_rpc_response({
-                "error": {
-                    "code": 1337,
-                    "message": "Error from RPC Server",
-                },
-            })
-            try:
-                rpc_client.add_shipment_data(STORAGE_CRED_ID, SHIPPER_WALLET_ID, CARRIER_WALLET_ID, '{}')
-                self.fail("Should have thrown RPC Error")
-            except RPCError as rpc_error:
-                self.assertEqual(rpc_error.status_code, 500)
-                self.assertEqual(rpc_error.detail, 'Error from RPC Server')
-
-            mock_method.return_value = mocked_rpc_response({
-                "result": {
-                    "code": 1337,
-                    "message": "Error from RPC Server",
-                },
-            })
-            try:
-                rpc_client.add_shipment_data(STORAGE_CRED_ID, SHIPPER_WALLET_ID, CARRIER_WALLET_ID, '{}')
-                self.fail("Should have thrown RPC Error")
-            except RPCError as rpc_error:
-                self.assertEqual(rpc_error.status_code, 500)
-                self.assertEqual(rpc_error.detail, 'Invalid response from Engine')
-
-            mock_method.return_value = mocked_rpc_response({
-                "jsonrpc": "2.0",
-                "result": {
-                    "success": True,
-                    "vault_signed": SHIPPER_WALLET_ID
-                },
-                "id": 0
-            })
-
-            vault_signed = rpc_client.add_shipment_data(STORAGE_CRED_ID, SHIPPER_WALLET_ID, CARRIER_WALLET_ID, '{}')
-            self.assertEqual(vault_signed, SHIPPER_WALLET_ID)
-
-    def test_add_tracking_data(self):
-
-        rpc_client = ShipmentRPCClient()
-
-        # Error response from RPC Server should return server detail in exception
-        with mock.patch.object(requests.Session, 'post') as mock_method:
-            mock_method.return_value = mocked_rpc_response({
-                "error": {
-                    "code": 1337,
-                    "message": "Error from RPC Server",
-                },
-            })
-            try:
-                rpc_client.add_tracking_data(STORAGE_CRED_ID, SHIPPER_WALLET_ID, CARRIER_WALLET_ID, '{}')
-                self.fail("Should have thrown RPC Error")
-            except RPCError as rpc_error:
-                self.assertEqual(rpc_error.status_code, 500)
-                self.assertEqual(rpc_error.detail, 'Error from RPC Server')
-
-            mock_method.return_value = mocked_rpc_response({
-                "result": {
-                    "code": 1337,
-                    "message": "Error from RPC Server",
-                },
-            })
-            try:
-                rpc_client.add_tracking_data(STORAGE_CRED_ID, SHIPPER_WALLET_ID, CARRIER_WALLET_ID, '{}')
-                self.fail("Should have thrown RPC Error")
-            except RPCError as rpc_error:
-                self.assertEqual(rpc_error.status_code, 500)
-                self.assertEqual(rpc_error.detail, 'Invalid response from Engine')
-
-            mock_method.return_value = mocked_rpc_response({
-                "jsonrpc": "2.0",
-                "result": {
-                    "success": True,
-                    "vault_signed": SHIPPER_WALLET_ID
-                },
-                "id": 0
-            })
-
-            vault_signed = rpc_client.add_tracking_data(STORAGE_CRED_ID, SHIPPER_WALLET_ID, CARRIER_WALLET_ID, '{}')
-            self.assertEqual(vault_signed, SHIPPER_WALLET_ID)
-
-    def test_create_shipment_transaction(self):
-
-        rpc_client = ShipmentRPCClient()
-
-        # Error response from RPC Server should return server detail in exception
-        with mock.patch.object(requests.Session, 'post') as mock_method:
-            mock_method.return_value = mocked_rpc_response({
-                "error": {
-                    "code": 1337,
-                    "message": "Error from RPC Server",
-                },
-            })
-            try:
-                rpc_client.create_shipment_transaction(STORAGE_CRED_ID, SHIPPER_WALLET_ID, 'SHIP', 400)
-                self.fail("Should have thrown RPC Error")
-            except RPCError as rpc_error:
-                self.assertEqual(rpc_error.status_code, 500)
-                self.assertEqual(rpc_error.detail, 'Error from RPC Server')
-
-            mock_method.return_value = mocked_rpc_response({
-                "result": {
-                    "code": 1337,
-                    "message": "Error from RPC Server",
-                },
-            })
-            try:
-                rpc_client.create_shipment_transaction(STORAGE_CRED_ID, SHIPPER_WALLET_ID, 'SHIP', 400)
-                self.fail("Should have thrown RPC Error")
-            except RPCError as rpc_error:
-                self.assertEqual(rpc_error.status_code, 500)
-                self.assertEqual(rpc_error.detail, 'Invalid response from Engine')
-
-            mock_method.return_value = mocked_rpc_response({
-                "jsonrpc": "2.0",
-                "result": {
-                    "success": True,
-                    "transaction": SHIPPER_WALLET_ID,
-                    "contractVersion": "v1"
-                },
-                "id": 0
-            })
-
-            contractVersion, transaction = rpc_client.create_shipment_transaction(STORAGE_CRED_ID, SHIPPER_WALLET_ID, CARRIER_WALLET_ID, '{}')
-            self.assertEqual(contractVersion, "v1")
-            self.assertEqual(transaction, SHIPPER_WALLET_ID)
-
-    def test_get_tracking_data(self):
-
-        rpc_client = ShipmentRPCClient()
-
-        # Error response from RPC Server should return server detail in exception
-        with mock.patch.object(requests.Session, 'post') as mock_method:
-            mock_method.return_value = mocked_rpc_response({
-                "error": {
-                    "code": 1337,
-                    "message": "Error from RPC Server",
-                },
-            })
-            try:
-                rpc_client.get_tracking_data(STORAGE_CRED_ID, SHIPPER_WALLET_ID, VAULT_ID)
-                self.fail("Should have thrown RPC Error")
-            except RPCError as rpc_error:
-                self.assertEqual(rpc_error.status_code, 500)
-                self.assertEqual(rpc_error.detail, 'Error from RPC Server')
-
-            mock_method.return_value = mocked_rpc_response({
-                "result": {
-                    "code": 1337,
-                    "message": "Error from RPC Server",
-                },
-            })
-            try:
-                rpc_client.get_tracking_data(STORAGE_CRED_ID, SHIPPER_WALLET_ID, VAULT_ID)
-                self.fail("Should have thrown RPC Error")
-            except RPCError as rpc_error:
-                self.assertEqual(rpc_error.status_code, 500)
-                self.assertEqual(rpc_error.detail, 'Invalid response from Engine')
-
-            mock_method.return_value = mocked_rpc_response({
-                "jsonrpc": "2.0",
-                "result": {
-                    "success": True,
-                    "contents": {
-                        "ayy": "lmao"
-                    }
-                },
-                "id": 0
-            })
-
-            contents = rpc_client.get_tracking_data(STORAGE_CRED_ID, SHIPPER_WALLET_ID, VAULT_ID)
-            self.assertEqual(contents, {
-                        "ayy": "lmao"
-                    })
-
-
-class TestLoad110RPCClient(APITestCase):
-    def test_set_vault_hash_tx(self):
-
-        rpc_client = Load110RPCClient()
-
-        # Error response from RPC Server should return server detail in exception
-        with mock.patch.object(requests.Session, 'post') as mock_method:
-            mock_method.return_value = mocked_rpc_response({
-                "error": {
-                    "code": 1337,
-                    "message": "Error from RPC Server",
-                },
-            })
-            try:
-                rpc_client.set_vault_hash_tx(SHIPPER_WALLET_ID, SHIPMENT_ID, VAULT_HASH)
-                self.fail("Should have thrown RPC Error")
-            except RPCError as rpc_error:
-                self.assertEqual(rpc_error.status_code, 500)
-                self.assertEqual(rpc_error.detail, 'Error from RPC Server')
-
-            mock_method.return_value = mocked_rpc_response({
-                "result": {
-                    "code": 1337,
-                    "message": "Invalid response from Engine",
-                },
-            })
-            try:
-                rpc_client.set_vault_hash_tx(SHIPPER_WALLET_ID, SHIPMENT_ID, VAULT_HASH)
-                self.fail("Should have thrown RPC Error")
-            except RPCError as rpc_error:
-                self.assertEqual(rpc_error.status_code, 500)
-                self.assertEqual(rpc_error.detail, 'Invalid response from Engine')
-
-            mock_method.return_value = mocked_rpc_response({
-                "jsonrpc": "2.0",
-                "result": {
-                    "success": True,
-                    "transaction": VAULT_HASH
-                },
-                "id": 0
-            })
-
-            transaction = rpc_client.set_vault_hash_tx(STORAGE_CRED_ID, SHIPPER_WALLET_ID, CARRIER_WALLET_ID)
-            self.assertEqual(transaction, VAULT_HASH)
-
-    def test_set_vault_uri_tx(self):
-
-        rpc_client = Load110RPCClient()
-
-        # Error response from RPC Server should return server detail in exception
-        with mock.patch.object(requests.Session, 'post') as mock_method:
-            mock_method.return_value = mocked_rpc_response({
-                "error": {
-                    "code": 1337,
-                    "message": "Error from RPC Server",
-                },
-            })
-            try:
-                rpc_client.set_vault_uri_tx(SHIPPER_WALLET_ID, SHIPMENT_ID, VAULT_ID)
-                self.fail("Should have thrown RPC Error")
-            except RPCError as rpc_error:
-                self.assertEqual(rpc_error.status_code, 500)
-                self.assertEqual(rpc_error.detail, 'Error from RPC Server')
-
-            mock_method.return_value = mocked_rpc_response({
-                "result": {
-                    "code": 1337,
-                    "message": "Invalid response from Engine",
-                },
-            })
-            try:
-                rpc_client.set_vault_uri_tx(SHIPPER_WALLET_ID, SHIPMENT_ID, VAULT_ID)
-                self.fail("Should have thrown RPC Error")
-            except RPCError as rpc_error:
-                self.assertEqual(rpc_error.status_code, 500)
-                self.assertEqual(rpc_error.detail, 'Invalid response from Engine')
-
-            mock_method.return_value = mocked_rpc_response({
-                "jsonrpc": "2.0",
-                "result": {
-                    "success": True,
-                    "transaction": VAULT_HASH
-                },
-                "id": 0
-            })
-
-            transaction = rpc_client.set_vault_uri_tx(STORAGE_CRED_ID, SHIPPER_WALLET_ID, VAULT_ID)
-            self.assertEqual(transaction, VAULT_HASH)
-
-    def test_set_carrier_tx(self):
-
-        rpc_client = Load110RPCClient()
-
-        # Error response from RPC Server should return server detail in exception
-        with mock.patch.object(requests.Session, 'post') as mock_method:
-            mock_method.return_value = mocked_rpc_response({
-                "error": {
-                    "code": 1337,
-                    "message": "Error from RPC Server",
-                },
-            })
-            try:
-                rpc_client.set_carrier_tx(SHIPPER_WALLET_ID, SHIPMENT_ID, CARRIER_WALLET_ID)
-                self.fail("Should have thrown RPC Error")
-            except RPCError as rpc_error:
-                self.assertEqual(rpc_error.status_code, 500)
-                self.assertEqual(rpc_error.detail, 'Error from RPC Server')
-
-            mock_method.return_value = mocked_rpc_response({
-                "result": {
-                    "code": 1337,
-                    "message": "Invalid response from Engine",
-                },
-            })
-            try:
-                rpc_client.set_carrier_tx(SHIPPER_WALLET_ID, SHIPMENT_ID, CARRIER_WALLET_ID)
-                self.fail("Should have thrown RPC Error")
-            except RPCError as rpc_error:
-                self.assertEqual(rpc_error.status_code, 500)
-                self.assertEqual(rpc_error.detail, 'Invalid response from Engine')
-
-            mock_method.return_value = mocked_rpc_response({
-                "jsonrpc": "2.0",
-                "result": {
-                    "success": True,
-                    "transaction": VAULT_HASH
-                },
-                "id": 0
-            })
-
-            transaction = rpc_client.set_carrier_tx(SHIPPER_WALLET_ID, SHIPMENT_ID, CARRIER_WALLET_ID)
-            self.assertEqual(transaction, VAULT_HASH)
-
-    def test_set_moderator_tx(self):
-
-        rpc_client = Load110RPCClient()
-
-        # Error response from RPC Server should return server detail in exception
-        with mock.patch.object(requests.Session, 'post') as mock_method:
-            mock_method.return_value = mocked_rpc_response({
-                "error": {
-                    "code": 1337,
-                    "message": "Error from RPC Server",
-                },
-            })
-            try:
-                rpc_client.set_moderator_tx(SHIPPER_WALLET_ID, SHIPMENT_ID, MODERATOR_WALLET_ID)
-                self.fail("Should have thrown RPC Error")
-            except RPCError as rpc_error:
-                self.assertEqual(rpc_error.status_code, 500)
-                self.assertEqual(rpc_error.detail, 'Error from RPC Server')
-
-            mock_method.return_value = mocked_rpc_response({
-                "result": {
-                    "code": 1337,
-                    "message": "Invalid response from Engine",
-                },
-            })
-            try:
-                rpc_client.set_moderator_tx(SHIPPER_WALLET_ID, SHIPMENT_ID, MODERATOR_WALLET_ID)
-                self.fail("Should have thrown RPC Error")
-            except RPCError as rpc_error:
-                self.assertEqual(rpc_error.status_code, 500)
-                self.assertEqual(rpc_error.detail, 'Invalid response from Engine')
-
-            mock_method.return_value = mocked_rpc_response({
-                "jsonrpc": "2.0",
-                "result": {
-                    "success": True,
-                    "transaction": VAULT_HASH
-                },
-                "id": 0
-            })
-
-            transaction = rpc_client.set_moderator_tx(SHIPPER_WALLET_ID, SHIPMENT_ID, MODERATOR_WALLET_ID)
-            self.assertEqual(transaction, VAULT_HASH)
